refactor(cli): inline command rendering (bd-unknown)
