+++ crates/beads-rs/src/cli/commands/deleted.rs
@@ -1,7 +1,7 @@
 use clap::Args;
 
-use super::super::render;
 use super::super::{Ctx, normalize_bead_id, print_ok, send};
+use super::fmt_wall_ms;
 use crate::api::QueryResult;
 use crate::daemon::ipc::{Request, ResponsePayload};
 use crate::{Error, Result};
@@ -47,14 +47,103 @@ pub(crate) fn handle(ctx: &Ctx, args: DeletedArgs) -> Result<()> {
 
     match ok {
         ResponsePayload::Query(QueryResult::Deleted(tombs)) => {
-            println!(
-                "{}",
-                render::render_deleted_list(&tombs, &args.since, args.all)
-            );
+            println!("{}", render_deleted_list(&tombs, &args.since, args.all));
             Ok(())
         }
         other => print_ok(&other, false),
     }
+}
+
+pub(crate) fn render_deleted_list(
+    tombs: &[crate::api::Tombstone],
+    since: &str,
+    all: bool,
+) -> String {
+    if tombs.is_empty() {
+        if all {
+            return "\n✨ No deletions tracked\n".into();
+        }
+        return format!("\n✨ No deletions in the last {since}\n");
+    }
+
+    let mut out = if all {
+        format!("\n🗑️ All tracked deletions ({} total):\n\n", tombs.len())
+    } else {
+        format!(
+            "\n🗑️ Deletions in the last {since} ({} total):\n\n",
+            tombs.len()
+        )
+    };
+
+    for t in tombs {
+        let ts = fmt_wall_ms(t.deleted_at.wall_ms);
+        let reason = t.reason.as_deref().unwrap_or("");
+        let reason = if reason.is_empty() {
+            "".to_string()
+        } else {
+            format!("  {reason}")
+        };
+        out.push_str(&format!(
+            "  {:<12}  {}  {:<12}{}\n",
+            t.id, ts, t.deleted_by, reason
+        ));
+    }
+
+    out.trim_end().into()
+}
+
+pub(crate) fn render_deleted(tombs: &[crate::api::Tombstone]) -> String {
+    if tombs.is_empty() {
+        return "\n✨ No deletions tracked\n".into();
+    }
+
+    let mut out = format!("\n🗑️ Deleted issues ({}):\n\n", tombs.len());
+    for t in tombs {
+        let ts = fmt_wall_ms(t.deleted_at.wall_ms);
+        let reason = t.reason.as_deref().unwrap_or("");
+        if reason.is_empty() {
+            out.push_str(&format!("  {:<12}  {}  {}\n", t.id, ts, t.deleted_by));
+        } else {
+            out.push_str(&format!(
+                "  {:<12}  {}  {}  {}\n",
+                t.id, ts, t.deleted_by, reason
+            ));
+        }
+    }
+    out.trim_end().into()
+}
+
+pub(crate) fn render_deleted_lookup(out: &crate::api::DeletedLookup) -> String {
+    if !out.found {
+        return format!(
+            "Issue {} not found in tombstones\n(This could mean the issue was never deleted, or the record was pruned)",
+            out.id
+        );
+    }
+    let record = match &out.record {
+        Some(r) => r,
+        None => {
+            return format!(
+                "Issue {} not found in tombstones\n(This could mean the issue was never deleted, or the record was pruned)",
+                out.id
+            );
+        }
+    };
+
+    let mut buf = format!("\n🗑️ Deletion record for {}:\n\n", out.id);
+    buf.push_str(&format!("  ID:        {}\n", record.id));
+    buf.push_str(&format!(
+        "  Deleted:   {}\n",
+        fmt_wall_ms(record.deleted_at.wall_ms)
+    ));
+    buf.push_str(&format!("  By:        {}\n", record.deleted_by));
+    if let Some(reason) = &record.reason
+        && !reason.is_empty()
+    {
+        buf.push_str(&format!("  Reason:    {}\n", reason));
+    }
+    buf.push('\n');
+    buf
 }
 
 fn parse_since_ms(s: &str) -> Result<u64> {
